moved jira services into subfolder
