+++ src/main/go/commands/user_view.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/sebastian-sommerfeld-io/jiracli/services/users"
+	"github.com/sebastian-sommerfeld-io/jiracli/services/jira/users"
 )
 
 type userViewOptions struct {
