+++ src/main/go/commands/user_view_test.go
@@ -3,7 +3,7 @@ package commands
 import (
 	"testing"
 
-	"github.com/sebastian-sommerfeld-io/jiracli/services/users"
+	"github.com/sebastian-sommerfeld-io/jiracli/services/jira/users"
 	"github.com/stretchr/testify/assert"
 )
