+++ src/main/go/services/jira/license/read_license.go
