+++ src/main/go/services/jira/users/count_users.go
