+++ src/main/go/services/jira/users/find_user.go
