+++ src/main/go/services/jira/users/find_user_test.go
