+++ src/main/go/services/jira/users/user_object.go
